+++ src/samples/randoms_batch/pool_item.rs
@@ -1,22 +1,30 @@
 use crate::{
     id_targeted::IdTargeted,
     pool_item::{new_pool_item_error::NewPoolItemError, PoolItem},
+    request_response::RequestResponse,
+    samples::Randoms,
+    sender_and_receiver::SenderAndReceiver,
     thread_request_response::*,
 };
+use std::fmt::Debug;
 
 use super::{randoms_batch_api::*, RandomsBatch};
 
-impl PoolItem for RandomsBatch {
-    type Init = RandomsBatchAddRequest;
+impl<P> PoolItem for RandomsBatch<P>
+where
+    P: SenderAndReceiver<Randoms> + Send + Sync + Debug,
+{
+    type Init = RandomsBatchAddRequest<P>;
     type Api = RandomsBatchApi;
 
     fn process_message(&mut self, request: Self::Api) -> ThreadRequestResponse<Self> {
         match request {
-            RandomsBatchApi::SumOfSums(request) => {
+            RandomsBatchApi::SumOfSums(RequestResponse::Request(request)) => {
                 let id = request.id();
                 let sum_of_sums = self.sum_of_sums();
                 SumOfSumsResponse { id, sum_of_sums }.into()
             }
+            _ => panic!("unexpected"),
         }
     }