+++ tests/example_two_level.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use messaging_thread_pool::{
     global_test_scope::global_test_scope, id_provider::id_provider_mutex::IdProviderMutex,
-    samples::*, thread_request_response::*, ThreadPool,
+    samples::*, sender_and_receiver::SenderAndReceiverMock, thread_request_response::*, ThreadPool,
 };
 use tracing::metadata::LevelFilter;
 
@@ -19,7 +19,7 @@ pub fn example_random_batches_() {
 
     // Create a thread pool for RandomsBatch
     // It is the lifetime of this struct that controls the lifetime of all of the pool items that are added
-    let randoms_batch_thread_pool = ThreadPool::<RandomsBatch>::new(2);
+    let randoms_batch_thread_pool = ThreadPool::<RandomsBatch<ThreadPool<Randoms>>>::new(2);
 
     // Create another thread pool to be used by the children of the RandomsBatches (which are Randoms)
     // The arrangement here is to have this thread shared by all of the Randoms regardless of which RandomsBatch
@@ -58,4 +58,35 @@ pub fn example_random_batches_() {
         .collect();
 
     dbg!(sum_of_sums);
+}
+
+#[test]
+pub fn example_random_batches_with_mock_thread_pool() {
+    // the id provider will not be used in this example but is required to construct a RandomsBatch
+    let id_provider = Arc::new(IdProviderMutex::new(0));
+
+    // Create a mock thread pool that will be called from inside of the RandomsBatch when sum-of_sums is called.
+    // Constructed like this is states that is expecting to receive 2 SumRequests requests (it will verify this)
+    // and in return it will return 2 SumResponses
+    // This enables the functionality of the sum_of_sums function to be tests without constructing a real thread pool
+    let randoms_thread_pool =
+        SenderAndReceiverMock::<Randoms, SumRequest, SumResponse>::new_with_expected_requests(
+            vec![SumRequest(2), SumRequest(4)],
+            vec![SumResponse { id: 2, sum: 2 }, SumResponse { id: 4, sum: 4 }],
+        );
+
+    // new create a RandomsBatch using the mock thread pool
+    let target = RandomsBatch {
+        id: 1,
+        contained_random_ids: vec![2, 4],
+        id_provider,
+        randoms_thread_pool: Arc::new(randoms_thread_pool),
+    };
+
+    // now call sum_of_sums
+    // this will fire the 2 expected requests and receive back the hard coded responses
+    let result = target.sum_of_sums();
+
+    // the sum of sums will be 6 (2 + 4)
+    assert_eq!(6, result);
 }